[IMPROVE] - Create separate directory for types and queries
