+++ graphql/mutations/user/createUserMutation.js
@@ -1,18 +1,16 @@
-import userInputType from '../types/userInputType.js';
-import db from '../../models/index.js';
-import userType from '../types/userType.js';
+import userInputType from '../../types/user/userInputType.js';
+import db from '../../../models/index.js';
+import userType from '../../types/user/userType.js';
 import bcrypt from 'bcrypt';
 
-const createUserMutationResolver = async (_, { user }, context) => {
+const createUserMutationResolver = async (_, {user}, context) => {
     const password = await bcrypt.hash(user.password, 5);
 
-    const createdUser = await db.User.create({
+    return await db.User.create({
         username: user.username,
         password: password,
     });
 
-    return createdUser;
-    
 }
 
 const createUserMutation = {
